requirements.txt
# altair==5.2.0
# annotated-types==0.6.0
# attrs==23.2.0
# beautifulsoup4==4.12.3
# blinker==1.7.0
# blis==0.7.11
# cachetools==5.3.2
# catalogue==2.0.10
# certifi==2024.2.2
# charset-normalizer==3.3.2
# click==8.1.7
# cloudpathlib==0.16.0
# confection==0.1.4
# contourpy==1.2.0
# cycler==0.12.1
# cymem==2.0.8
# entrypoints==0.4
faiss-cpu==1.7.4
# Faker==22.7.0
# favicon==0.7.0
# fonttools==4.47.2
# gitdb==4.0.11
# GitPython==3.1.41
# htbuilder==0.6.2
# idna==3.6
# importlib-metadata==7.0.1
# Jinja2==3.1.3
# joblib==1.3.2
# jsonschema==4.21.1
# jsonschema-specifications==2023.12.1
# kiwisolver==1.4.5
# langcodes==3.3.0
# lxml==5.1.0
# Markdown==3.5.2
# markdown-it-py==3.0.0
# markdownlit==0.0.7
# MarkupSafe==2.1.5
#matplotlib==3.8.2
# mdurl==0.1.2
# more-itertools==10.2.0
# murmurhash==1.0.10
nltk==3.8.1
numpy==1.26.4
# packaging==23.2
pandas==2.2.0
# patsy==0.5.6
# pillow==10.2.0
plotly==5.18.0
plotly-express==0.4.1
preshed==3.0.9
# prometheus-client==0.19.0
# protobuf==4.25.2
# pyarrow==15.0.0
# pydantic==2.6.1
# pydantic_core==2.16.2
# pydeck==0.8.1b0
# Pygments==2.17.2
# pymdown-extensions==10.7
# pyparsing==3.1.1
# python-dateutil==2.8.2
# pytz==2024.1
# PyYAML==6.0.1
# referencing==0.33.0
# regex==2023.12.25
# requests==2.31.0
# rich==13.7.0
# rpds-py==0.17.1
 scipy==1.12.0
# six==1.16.0
# smart-open==6.4.0
# smmap==5.0.1
# soupsieve==2.5
spacy==3.7.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
st-annotated-text==4.0.1
statsmodels==0.14.1
streamlit==1.31.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-embedcode==0.1.2
streamlit-extras==0.4.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.2
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
# tenacity==8.2.3
# thinc==8.2.2
# toml==0.10.2
# toolz==0.12.1
# tornado==6.4
# tqdm==4.66.1
# typer==0.9.0
# typing_extensions==4.9.0
# tzdata==2023.4
# tzlocal==5.2
# urllib3==2.2.0
# validators==0.22.0
# wasabi==1.1.2
# weasel==0.3.4
# zipp==3.17.0
